backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
faiss-cpu
sentence-transformers
numpy
groq
python-dotenv
PyMuPDF
langchain
requests
